requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.40
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
psycopg2==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
scipy==1.14.0
Werkzeug==3.0.3